pyproject.toml
[project]
name = "ai-tg-bot"
version = "0.1.0"
description = "AI-powered Telegram bot with streaming responses and tool support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.0",
    "openai>=1.60.0",
    "sqlalchemy>=2.0.46",
    "aiosqlite>=0.22.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.0",
    "tavily-python>=0.7.0",
    "tiktoken>=0.12.0",
]
